pkg/rbfs/commons/context.go
/*
 * Copyright (C) 2021, RtBrick, Inc.
 * SPDX-License-Identifier: BSD-3-Clause
 */

package commons

import (
	"context"
	"fmt"
	"net/url"

	"github.com/rtbrick/go-rbfs-client/pkg/rbfs/state"
)

const (
	ctrldURLKey          = contextKey("CtrldURL")
	elementNameKey       = contextKey("ElementName")
	OpsdServiceName      = ServiceName("opsd")
	RestconfdServiceName = ServiceName("restconfd")
)

type (
	contextKey string

	ServiceName string

	// RbfsContext provides access to all request-specific settings to invoke the RBFS REST API.
	// It is also a go context that allows cancelling a request.
	RbfsContext interface {
		context.Context

		// GetServiceEndpoint computes the REST API endpoint for the given service.
		GetServiceEndpoint(ServiceName) (*url.URL, error)
	}

	rbfsContext struct {
		context.Context
	}

	// RbfsContextOption allows applying an optional RBFS context setting.
	RbfsContextOption func(ctx context.Context) (context.Context, error)
)

// RbfsAccessToken adds an access token the a RBFS context.
func RbfsAccessToken(token string) RbfsContextOption {
	return func(ctx context.Context) (context.Context, error) {
		if token != "" {
			// Populate access token to all APIs
			ctx = context.WithValue(ctx, state.ContextAccessToken, token)
		}
		return ctx, nil
	}
}

// NewRbfsContext creates a new RBFS context from the given context to access an RBFS instance available under the
// given endpointURL with the specified elementName and the request options.
func NewRbfsContext(ctx context.Context, endpointURL *url.URL, elementName string, options ...RbfsContextOption) (*rbfsContext, error) {
	ctx = context.WithValue(ctx, ctrldURLKey, endpointURL)
	ctx = context.WithValue(ctx, elementNameKey, elementName)
	var err error
	for _, option := range options {
		// Apply optional settings to the diagnostic context
		ctx, err = option(ctx)
		if err != nil {
			return nil, err
		}
	}
	return &rbfsContext{Context: ctx}, nil
}

// MustRbfsContext creates a new RBFS context from the given context.
func MustRbfsContext(ctx context.Context) *rbfsContext {
	_, ok := ctx.Value(ctrldURLKey).(*url.URL)
	if !ok {
		panic("ctrldEndpoint not set")
	}
	_, ok = ctx.Value(elementNameKey).(string)
	if !ok {
		panic("elementName not set")
	}
	return &rbfsContext{Context: ctx}
}

func (r *rbfsContext) GetServiceEndpoint(serviceName ServiceName) (*url.URL, error) {
	if serviceName == "" {
		return nil, fmt.Errorf("empty service name is not supported")
	}
	ctrldEndpoint := r.Value(ctrldURLKey).(*url.URL)
	elementName := r.Value(elementNameKey).(string)
	serviceEndpoint := fmt.Sprintf("%v/api/v1/rbfs/elements/%v/services/%v/proxy", ctrldEndpoint, elementName, serviceName)

	return url.Parse(serviceEndpoint)
}
